@@ -1,6 +1,6 @@
 {
   "name": "ngx.leaflet.components",
-  "version": "1.4.2",
+  "version": "2.0.0",
   "description": "an angular.io component collection for leaflet",
   "main": "./bundles/src/ngx.leaflet.components.umd.js",
   "jsnext:main": "./esm5/src/ngx.leaflet.components.js",
@@ -42,30 +42,29 @@
     "url": "https://github.com/elasticrash/ngx.leaflet.component/issues"
   },
   "peerDependencies": {
-    "tslib": "^1.7.1",
-    "@angular/core": ">=4.0.0",
-    "@angular/http": ">=4.0.0",
-    "@angular/common": ">=4.0.0",
-    "rxjs": ">=5.0.0",
+    "tslib": "1.9.2",
+    "@angular/core": ">=6.0.0",
+    "@angular/http": ">=6.0.0",
+    "@angular/common": ">=6.0.0",
+    "rxjs": ">=6.0.0",
     "leaflet": ">=1.0.0",
-    "typescript": ">=2.4.2"
+    "typescript": ">=2.7.2"
   },
   "homepage": "https://github.com/elasticrash/ngx.leaflet.component#readme",
   "devDependencies": {
-    "@angular/common": "5.0.2",
-    "@angular/compiler": "5.0.2",
-    "@angular/compiler-cli": "5.0.2",
-    "@angular/core": "5.0.2",
-    "@angular/http": "5.0.2",
-    "@angular/platform-browser": "5.0.2",
-    "@angular/platform-browser-dynamic": "5.0.2",
-    "@compodoc/compodoc": "^1.0.0-beta.1",
-    "@types/geojson": "^1.0.2",
-    "@types/jasmine": "^2.5.53",
-    "@types/leaflet": "^1.2.0",
-    "@types/node": "^8.0.28",
+    "@angular/common": "6.0.5",
+    "@angular/compiler": "6.0.5",
+    "@angular/compiler-cli": "6.0.5",
+    "@angular/core": "6.0.5",
+    "@angular/http": "6.0.5",
+    "@angular/platform-browser": "6.0.5",
+    "@angular/platform-browser-dynamic": "6.0.5",
+    "@compodoc/compodoc": "1.1.3",
+    "@types/geojson": "7946.0.3",
+    "@types/jasmine": "2.8.8",
+    "@types/leaflet": "1.2.7",
     "angular2-template-loader": "https://github.com/elasticrash/angular2-template-loader.git",
-    "awesome-typescript-loader": "^3.2.3",
+    "awesome-typescript-loader": "5.0.0",
     "chalk": "2.3.0",
     "codelyzer": "4.0.0",
     "core-js": "^2.5.0",
@@ -101,17 +100,17 @@
     "rollup-plugin-license": "0.5.0",
     "rollup-plugin-node-resolve": "3.0.0",
     "rollup-plugin-sourcemaps": "0.4.2",
-    "rxjs": "5.5.2",
+    "rxjs": "6.2.1",
     "shelljs": "0.7.8",
     "source-map-loader": "0.2.3",
     "source-map-support": "^0.4.15",
     "ts-loader": "3.1.1",
     "ts-node": "1.2.1",
-    "tslint": "^5.8.0",
+    "tslint": "^5.10.0",
     "typedoc": "^0.8.0",
-    "typescript": "2.4.2",
+    "typescript": "2.7.2",
     "uglify-js": "3.1.6",
     "webpack": "^3.8.1",
-    "zone.js": "^0.8.18"
+    "zone.js": "^0.8.26"
   }
 }
