@@ -1,14 +1,13 @@
-import { Component, Input, Optional, ElementRef, ViewChild } from '@angular/core';
-import { Http, Response, Headers, RequestOptions, Request, RequestMethod, ResponseContentType } from '@angular/http';
+import { Component, Input, Optional, ElementRef, ViewChild, OnInit } from '@angular/core';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { MapService } from '../services/map.service';
 import { GroupService } from '../services/group.service';
 import { PopupService } from '../services/popup.service';
 import { LeafletElement } from '../map/map';
 import { LeafletGroup } from '../group/group';
 import { CoordinateHandler } from '../helpers/coordinateHandler';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import * as L from 'leaflet';
 
 @Component({
@@ -18,107 +17,105 @@ import * as L from 'leaflet';
   providers: [PopupService]
 })
 
-export class MarkerElement extends CoordinateHandler {
-  @Input() lat: number = 52.6;
-  @Input() lon: number = -1.1;
-  @Input() mouseover: string | undefined = undefined;
-  @Input() onclick: string | undefined = undefined;
-  @Input() iconUrl: string = "";
-  @ViewChild('ngel') ngEl: ElementRef;  
+export class MarkerElement extends CoordinateHandler implements OnInit {
+  @Input() public lat: number = 52.6;
+  @Input() public lon: number = -1.1;
+  @Input() public mouseover: string | undefined = undefined;
+  @Input() public onclick: string | undefined = undefined;
+  @Input() public iconUrl: string = "";
+  @ViewChild('ngel') public ngEl: ElementRef;
   public marker: any = null;
 
   constructor(
     private mapService: MapService,
     private popupService: PopupService,
-    private http: Http,
-    @Optional() private groupService?: GroupService,        
-    @Optional() private LeafletElement?: LeafletElement,
-    @Optional() private LeafletGroup?: LeafletGroup) {
+    private http: HttpClient,
+    @Optional() private groupService?: GroupService,
+    @Optional() private leafletElement?: LeafletElement,
+    @Optional() private leafletGroup?: LeafletGroup) {
     super();
   }
 
-  ngOnInit() {
+  public ngOnInit() {
     super.assignCartesianPointToLeafletsLatLngSchema();
-    var model = this;
-    if (this.LeafletElement || this.LeafletGroup) {
+    const model = this;
+    if (this.leafletElement || this.leafletGroup) {
 
-      let map = this.mapService.getMap();
+      const mapObject = this.mapService.getMap();
 
-      super.transformPointCoordinates(this.LeafletElement.crs);
+      super.transformPointCoordinates(this.leafletElement.crs);
 
       if (this.iconUrl === "") {
         this.marker = L.marker([this.lat, this.lon]);
-        this.createMarkerlayer(this.marker, map);
+        this.createMarkerlayer(this.marker, mapObject);
       } else {
-        this.imageExists(this.iconUrl, function (exists) {
+        this.imageExists(this.iconUrl, (exists) => {
 
           model.getImage().subscribe(
-            image => {
-              var img = document.createElement("img");
+            (image) => {
+              const img = document.createElement("img");
               window.URL.createObjectURL(image.blob());
-              var reader = new FileReader();
-              reader.onload = function () {
+              const reader = new FileReader();
+              reader.onload = () => {
                 img.src = reader.result;
-                var myIcon = L.icon({
+                const myIcon = L.icon({
                   iconUrl: model.iconUrl,
                   iconSize: [img.width, img.height],
                   iconAnchor: [img.width / 2, img.height - 1],
                   popupAnchor: [0, -img.height]
                 });
 
-                var obj = { icon: myIcon, options: null };
+                const obj = { icon: myIcon, options: null };
                 model.marker = L.marker([model.lat, model.lon], obj);
-                model.createMarkerlayer(model.marker, map);
-              }
+                model.createMarkerlayer(model.marker, mapObject);
+              };
               reader.readAsDataURL(image.blob());
-            },
-            err => {
+            }, (err) => {
+              // tslint:disable-next-line:no-console
               console.log(err);
             });
         });
       }
     } else {
-      console.warn("This marker-element will not be rendered \n the expected parent node of marker-element should be either leaf-element or leaflet-group");
+      // tslint:disable-next-line:no-console
+      console.warn(`This marker-element will not be rendered
+      the expected parent node of marker-element should be either leaf-element or leaflet-group`);
     }
   }
 
-  createMarkerlayer(marker, map) {
-    var textInput = undefined;
+  public createMarkerlayer(marker, mapObject) {
+    let textInput;
     if (this.ngEl.nativeElement.childNodes.length > 0) {
-      var textNode = this.ngEl.nativeElement.childNodes[0];
+      const textNode = this.ngEl.nativeElement.childNodes[0];
       textInput = textNode.nodeValue;
     }
 
-    //add popup methods on element only if any of the tests are not undefined
+    // add popup methods on element only if any of the tests are not undefined
     if (this.mouseover !== undefined || this.onclick !== undefined || textInput !== undefined) {
       this.popupService.enablePopup(this.mouseover, this.onclick, this.marker, textInput);
     }
 
-    //only if the parent is map should the marker-element should be directly added to the map
-    if (this.LeafletGroup) {
-      this.groupService.addOLayersToGroup(marker, map, this.mapService, this.LeafletGroup);
+    // only if the parent is map should the marker-element should be directly added to the map
+    if (this.leafletGroup) {
+      this.groupService.addOLayersToGroup(marker, mapObject, this.mapService, this.leafletGroup);
     } else {
-      marker.addTo(map);
+      marker.addTo(mapObject);
     }
   }
 
-  imageExists(url, callback) {
-    var img = new Image();
-    img.onload = function () { callback(true); };
-    img.onerror = function () { callback(false); };
+  public imageExists(url, callback) {
+    const img = new Image();
+    img.onload = () => { callback(true); };
+    img.onerror = () => { callback(false); };
     img.src = url;
   }
 
-  getImage(): any {
-    let headers = new Headers({ 'Content-Type': 'application/x-www-form-urlencoded' });
-    let options = new RequestOptions({
-      responseType: ResponseContentType.Blob,
-      headers: headers
-    });
-
-
-    return this.http.get(this.iconUrl, options)
-    .map((res: Response) => res)
-    .catch((error: any) => Observable.throw('Server error'));
+  public getImage(): any {
+    const headers = new HttpHeaders().set('Content-Type', 'application/x-www-form-urlencoded');
+    return this.http.get(this.iconUrl, {
+      headers,
+      responseType: "blob"
+    }).pipe(map((res: Response) => res)).toPromise()
+      .catch((error: any) => Observable.throw('Server error'));
   }
 }
