@@ -2,18 +2,16 @@ import { Component, Input } from '@angular/core';
 import { MapService } from '../services/map.service';
 import { GroupService } from '../services/group.service';
 import { GuidService } from '../services/globalId.service';
-import * as L from 'leaflet';
-
 
 @Component({
     selector: 'leaflet-group',
     template: ``,
     styles: [''],
-    providers: [GroupService]    
+    providers: [GroupService]
 })
 
 export class LeafletGroup {
-    @Input() name: string = '';
+    @Input() public name: string = '';
     public globalId: string = this.guidService.newGuid();
 
     constructor(
@@ -21,12 +19,4 @@ export class LeafletGroup {
         private groupService: GroupService,
         private guidService: GuidService) {
     }
-
-    ngOnInit() {
-    }
-
-    ngAfterViewInit() {
-    }
-
-
 }
