@@ -1,12 +1,9 @@
 import { Input } from '@angular/core';
 
 export class GeoJSONCoordinateHandler {
-    @Input() geojson: any = {};
+    @Input() public geojson: any = {};
 
-    constructor() {
-    }
-
-    transformJSONCoordinates(geoJSON: any, crs: any) {
+    public transformJSONCoordinates(geoJSON: any, crs: any) {
         /**
          * 7.  GeoJSON Types Are Not Extensible
          * Implementations MUST NOT extend the fixed set of GeoJSON types:
@@ -14,66 +11,66 @@ export class GeoJSONCoordinateHandler {
          * MultiLineString, MultiPolygon, and GeometryCollection.
          */
         if (geoJSON.type === "FeatureCollection") {
-            var featureCollection = geoJSON;
-            featureCollection.features.forEach(feature => {
+            const featureCollection = geoJSON;
+            featureCollection.features.forEach((feature) => {
                 this.transformJSONCoordinates(feature, crs);
             });
         }
         if (geoJSON.type === "Feature") {
-            var feature = geoJSON;
+            const feature = geoJSON;
             this.transformJSONCoordinates(feature.geometry, crs);
         }
         if (geoJSON.type === "Point") {
-            var point = geoJSON;
-            point = this.transformPointCoordinates(point.coordinates, crs);
+            const point = geoJSON;
+            this.transformPointCoordinates(point.coordinates, crs);
         }
         if (geoJSON.type === "LineString") {
-            var lineString = geoJSON;
-            lineString.coordinates.forEach(point => {
+            const lineString = geoJSON;
+            lineString.coordinates.forEach((point) => {
                 this.transformPointCoordinates(point, crs);
             });
         }
         if (geoJSON.type === "MultiPoint") {
-            var multiPoint = geoJSON;
-            multiPoint.coordinates.forEach(point => {
+            const multiPoint = geoJSON;
+            multiPoint.coordinates.forEach((point) => {
                 this.transformPointCoordinates(point, crs);
             });
         }
         if (geoJSON.type === "Polygon") {
-            var polygon = geoJSON;
-            polygon.coordinates.forEach(polygonElement => {
-                polygonElement.forEach(point => {
+            const polygon = geoJSON;
+            polygon.coordinates.forEach((polygonElement) => {
+                polygonElement.forEach((point) => {
                     this.transformPointCoordinates(point, crs);
                 });
             });
         }
         if (geoJSON.type === "MultiLineString") {
-            var multiLineString = geoJSON;
-            multiLineString.coordinates.forEach(lineString => {
-                lineString.forEach(point => {
+            const multiLineString = geoJSON;
+            multiLineString.coordinates.forEach((lineString) => {
+                lineString.forEach((point) => {
                     this.transformPointCoordinates(point, crs);
                 });
             });
         }
         if (geoJSON.type === "MultiPolygon") {
-            var multiPolygon = geoJSON;
-            multiPolygon.coordinates.forEach(polygon => {
-                polygon.forEach(polygonElement => {
-                    polygonElement.forEach(point => {
+            const multiPolygon = geoJSON;
+            multiPolygon.coordinates.forEach((polygon) => {
+                polygon.forEach((polygonElement) => {
+                    polygonElement.forEach((point) => {
                         this.transformPointCoordinates(point, crs);
                     });
                 });
             });
         }
         if (geoJSON.type === "GeometryCollection") {
-            var geometryCollection = geoJSON;
-            geometryCollection.geometries.forEach(geometry => {
+            const geometryCollection = geoJSON;
+            geometryCollection.geometries.forEach((geometry) => {
                 this.transformJSONCoordinates(geometry, crs);
             });
         }
     }
 
-    transformPointCoordinates(point, crs) {
+    public transformPointCoordinates(point, crs) {
         /**
          * this is because leaflet default CRS is 3857 (so it can render wms properly)
          * but uses 4326 everywhere else so if CRS is 3857 don't reproject coordinates
@@ -82,7 +79,7 @@ export class GeoJSONCoordinateHandler {
          * how leaflet doesn't handle projections on a global state
          */
         if (crs.code && crs.code !== "EPSG:3857") {
-            let newlatlng = crs.unproject({ x: point[0], y: point[1] });
+            const newlatlng = crs.unproject({ x: point[0], y: point[1] });
             point[1] = newlatlng.lat;
             point[0] = newlatlng.lng;
             return point;
