@@ -1,5 +1,5 @@
-import { LeafletElement } from './map/map'
-import { AttributionControl } from './map/attribution/attribution'
+import { LeafletElement } from './map/map';
+import { AttributionControl } from './map/attribution/attribution';
 import { ScaleControl } from './map/scale/scale';
 import { ZoomControl } from './map/zoom/zoom';
 import { WatermarkControl } from './map/watermark/watermark';
@@ -16,20 +16,20 @@ import { GeoJsonElement } from './geojson/geojson';
 import { PopupElement } from './popup/popup';
 import { LeafletGroup } from './group/group';
 
-import { MapService } from './services/map.service'
-import { GroupService } from './services/group.service'
+import { MapService } from './services/map.service';
+import { GroupService } from './services/group.service';
 import { PopupService } from './services/popup.service';
 import { GuidService } from './services/globalId.service';
 import { HelperService } from './services/helper.service';
 
 export const CandTLeafletComponent = [
-    //map and controls
+    // map and controls
     LeafletElement,
     AttributionControl,
     ScaleControl,
     ZoomControl,
     WatermarkControl,
-    //layers and vectors
+    // layers and vectors
     LayerElement,
     ImageOverlayElement,
     MarkerElement,
@@ -39,7 +39,7 @@ export const CandTLeafletComponent = [
     PolylineElement,
     GeoJsonElement,
     PopupElement,
-    //rest
+    // rest
     LeafletGroup
 ];
 
