@@ -15,13 +15,13 @@ export { PolylineElement } from './polyline/polyline';
 export { GeoJsonElement } from './geojson/geojson';
 export { PopupElement } from './popup/popup';
 export { LeafletGroup } from './group/group';
-export { MapService } from './services/map.service'
-export { GroupService } from './services/group.service'
+export { MapService } from './services/map.service';
+export { GroupService } from './services/group.service';
 export { PopupService } from './services/popup.service';
 export { GuidService } from './services/globalId.service';
 export { HelperService } from './services/helper.service';
-export { attributionModel } from './models/attributionModel';
-export { path } from './models/path';
-export { scaleModel } from './models/scaleModel';
-export { zoomModel } from './models/zoomModel';
+export { AttributionModel } from './models/AttributionModel';
+export { Path } from './models/path';
+export { ScaleModel } from './models/scaleModel';
+export { ΖoomModel } from './models/zoomModel';
 export { ngxLeafletModule } from './ngx.leaflet.module';
