@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
-import { LeafletElement } from './map/map'
-import { AttributionControl } from './map/attribution/attribution'
+import { LeafletElement } from './map/map';
+import { AttributionControl } from './map/attribution/attribution';
 import { ScaleControl } from './map/scale/scale';
 import { ZoomControl } from './map/zoom/zoom';
 import { WatermarkControl } from './map/watermark/watermark';
@@ -17,8 +17,8 @@ import { GeoJsonElement } from './geojson/geojson';
 import { PopupElement } from './popup/popup';
 import { LeafletGroup } from './group/group';
 
-import { MapService } from './services/map.service'
-import { GroupService } from './services/group.service'
+import { MapService } from './services/map.service';
+import { GroupService } from './services/group.service';
 import { PopupService } from './services/popup.service';
 import { GuidService } from './services/globalId.service';
 import { HelperService } from './services/helper.service';
@@ -52,7 +52,7 @@ import { CommonModule } from "@angular/common";
         PopupService,
         GuidService,
         HelperService
-        ],
+    ],
     exports: [
         LeafletElement,
         AttributionControl,
