@@ -1,1 +1,1 @@
-export * from './polygon'; 
+export * from './polygon';
